Fix compilation warnings

+++ src/attestUtxo.c
@@ -329,7 +329,7 @@ void handle_attestUtxo(
         uint8_t p1,
         uint8_t p2,
         uint8_t* dataBuffer,
-        uint16_t dataLength
+        size_t dataLength
 )
 {
 	attestUtxoState_t* state = &global_state;
