+++ src/attestUtxo.h
@@ -97,7 +97,7 @@ void handle_attestUtxo(
         uint8_t p1,
         uint8_t p2,
         uint8_t* dataBuffer,
-        uint16_t dataLength
+        size_t dataLength
 );
 
 void run_test_attestUtxo();
